+++ libs/dot-components/src/lib/components/split-button/SplitButton.styles.ts
@@ -1,0 +1,12 @@
+import { ButtonGroup } from '@material-ui/core';
+import styled, { css } from 'styled-components';
+
+export const rootClassName = 'dot-split-button-group';
+
+export const StyledSplitButtonGroup = styled(ButtonGroup)`
+  &.${rootClassName} {
+    .dot-button {
+      margin: 0;
+    }
+  }
+`;